+++ Readme.md
@@ -1,12 +1,5 @@
 # TODO
 
-- [x] Create a lambda function called getProductsList
-- [x] Create a lambda function called getProductsById
-- [x] Swagger documentation is created for Product Service
-- [x] Lambda handlers are covered by basic UNIT tests
-- [x] Lambda handlers (getProductsList, getProductsById) code is written not in 1 single module
-- [x] Main error scenarios are handled by API
-
 # Commands
 
 - npm run cdk:deploy - deploy functions
