--- cdk.out/cdk.out
@@ -1,1 +1,0 @@
-{"version":"35.0.0"}
