+++ productService/jest.config.ts
@@ -1,0 +1,12 @@
+export default {
+  preset: "ts-jest",
+  transform: {
+    "^.+\\.tsx?$": "ts-jest",
+  },
+  testEnvironment: "node",
+  testPathIgnorePatterns: ["/node_modules/", "/dist/"],
+  moduleNameMapper: {
+    "^@/(.*)$": "<rootDir>/src/$1",
+    "\\.(css|less|scss)$": "identity-obj-proxy",
+  },
+};
