+++ productService/package.json
@@ -1,0 +1,38 @@
+{
+  "name": "nodejs-aws-shop-backend",
+  "version": "1.0.0",
+  "description": "",
+  "type": "module",
+  "main": "index.js",
+  "scripts": {
+    "test": "jest",
+    "cdk:deploy": "cdk deploy"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/andreypotkas/nodejs-aws-shop-backend.git"
+  },
+  "author": "",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/andreypotkas/nodejs-aws-shop-backend/issues"
+  },
+  "homepage": "https://github.com/andreypotkas/nodejs-aws-shop-backend#readme",
+  "dependencies": {
+    "@aws-cdk/aws-apigatewayv2-alpha": "^2.110.0-alpha.0",
+    "@aws-cdk/aws-apigatewayv2-integrations-alpha": "^2.110.0-alpha.0",
+    "aws-cdk-lib": "^2.110.0",
+    "aws-sdk": "^2.1502.0",
+    "dotenv": "^16.3.1",
+    "typescript": "^5.3.2",
+    "uuid": "^9.0.1"
+  },
+  "devDependencies": {
+    "@types/jest": "^29.5.10",
+    "@types/uuid": "^9.0.7",
+    "esbuild": "^0.19.5",
+    "jest": "^29.7.0",
+    "ts-jest": "^29.1.1",
+    "ts-node": "^10.9.1"
+  }
+}
