+++ productService/src/data/data.ts
@@ -1,0 +1,62 @@
+export const dataArray = [
+  {
+    id: "1",
+    title: "iPhone 13",
+    description: "Powerful and sleek",
+    price: 799.99,
+  },
+  {
+    id: "2",
+    title: "iPhone 13 Pro",
+    description: "Pro-level performance",
+    price: 999.99,
+  },
+  {
+    id: "3",
+    title: "iPhone SE",
+    description: "Compact and affordable",
+    price: 399.99,
+  },
+  {
+    id: "4",
+    title: "iPhone 12 Mini",
+    description: "Small size, big power",
+    price: 699.99,
+  },
+  {
+    id: "5",
+    title: "iPhone 12",
+    description: "Brilliant display, powerful camera",
+    price: 899.99,
+  },
+  {
+    id: "6",
+    title: "iPhone 11",
+    description: "Reliable and feature-packed",
+    price: 699.99,
+  },
+  {
+    id: "7",
+    title: "iPhone XR",
+    description: "Great value for performance",
+    price: 499.99,
+  },
+  {
+    id: "8",
+    title: "iPhone XS",
+    description: "Premium design and features",
+    price: 799.99,
+  },
+  {
+    id: "9",
+    title: "iPhone X",
+    description: "Revolutionary design",
+    price: 899.99,
+  },
+  {
+    id: "10",
+    title: "iPhone 8",
+    description: "Classic design, modern features",
+    price: 599.99,
+  },
+];
